requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.5
chardet==4.0.0
discord.py==1.7.2
idna==3.1
multidict==5.1.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.17.1
six==1.16.0
typing-extensions==3.10.0.0
yarl==1.6.3
ffmpeg==1.4